% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_documento_stj.R
\name{baixar_documento_stj}
\alias{baixar_documento_stj}
\title{Baixar documentos do processo, ex. decisões, certidões etc}
\usage{
baixar_documento_stj(sequencial, plano = "sequential", diretorio = ".",
  formato = c("PDF", "HTML"))
}
\arguments{
\item{sequencial}{códigos dos documentos obtido com ler_metadocs_stj}

\item{plano}{ver `future::plan`}

\item{diretorio}{se não informado, diretório atual}

\item{formato}{PDF ou HTML}
}
\description{
Baixar documentos do processo, ex. decisões, certidões etc
}
\examples{
\dontrun{
baixar_documento_stj("1764491")
}
}
