% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_pdf_stj.R
\name{baixar_pdfs_stj}
\alias{baixar_pdfs_stj}
\title{Baixar decisões do processo}
\usage{
baixar_pdfs_stj(sequencial, plano = "sequential", diretorio = ".")
}
\arguments{
\item{sequencial}{códigos dos documentos}

\item{plano}{ver `future::plan`}

\item{diretorio}{se não informado, diretório atual}
}
\description{
Baixar decisões do processo
}
\examples{
\dontrun{
baixar_pdf_stj("1764491")
}
}
