% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_processo_stj.R
\name{baixar_processo_stj}
\alias{baixar_processo_stj}
\title{Baixa processos do stj pelo número do registro}
\usage{
baixar_processo_stj(registro_stj = NULL, diretorio = ".")
}
\arguments{
\item{registro_stj}{Número do registro}

\item{diretorio}{Diretório. Default para atual}
}
\value{
html baixado no diretório indicado
}
\description{
Baixa processos do stj pelo número do registro
}
\examples{
\dontrun{
baxiar_processo_stj("2016/0074331-0")

}
}
