% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificar_recurso.R
\name{classificar_recurso}
\alias{classificar_recurso}
\title{Classifica recursos}
\usage{
classificar_recurso(x, dispositivo, decisao)
}
\arguments{
\item{x}{data.frame com a coluna a ser classificada}

\item{dispositivo}{Coluna onde se encontram os dispositivos do acc처rd찾o}

\item{decisao}{Nome da coluna a ser criada com a decis찾o}
}
\value{
x adicionado da coluna decis찾o.
}
\description{
Classifica recursos
}
