% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj_comunicacao.R
\name{cnj_consultar_comunicacao1}
\alias{cnj_consultar_comunicacao1}
\title{Baixa comunicações do STJ via api do CNJ}
\usage{
cnj_consultar_comunicacao1(
  processo = NULL,
  dt_inicio = NULL,
  dt_fim = NULL,
  diretorio = "."
)
}
\arguments{
\item{processo}{Número único do processo}

\item{dt_inicio}{Data do início das comunicações}

\item{dt_fim}{data do fim das comunicações}

\item{diretorio}{Data final}
}
\value{
json
}
\description{
Baixa comunicações do STJ via api do CNJ
}
\details{
Tente restringir a data_inicio e data_fim para
   o mesmo dia ou alguns poucos dias.
}
