% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj_comunicacao.R
\name{cnj_ler_comunicacao}
\alias{cnj_ler_comunicacao}
\title{Lê comunicação do CNJ}
\usage{
cnj_ler_comunicacao(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Se arquivos não foi fornecido, tenta encontrar arquivos no diretório}
}
\value{
tibble
}
\description{
Lê comunicação do CNJ
}
