% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnj_comunicacao.R
\name{cnj_parsear_texto}
\alias{cnj_parsear_texto}
\title{Parsea coluna texto}
\usage{
cnj_parsear_texto(x)
}
\arguments{
\item{x}{coluna texto em html}
}
\value{
texto
}
\description{
Parsea coluna texto
}
