% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_detalhes_stj.R
\name{ler_detalhes_stj}
\alias{ler_detalhes_stj}
\title{Ler detalhes dos processos do STJ}
\usage{
ler_detalhes_stj(diretorio = ".", arquivos = NULL)
}
\arguments{
\item{diretorio}{Diretório se arquivos não forem informados}

\item{arquivos}{Se os caminhos para os arquivos forem fornecidos
o diretório é ignorado.}
}
\value{
tibble com detalhes do processo
}
\description{
Ler detalhes dos processos do STJ
}
\examples{
\dontrun{
df <- ler_detalhes_stj(diretorio = ".")
}
}
