% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_documento_stj.R
\name{ler_documento_stj}
\alias{ler_documento_stj}
\title{Ler documentos do STJ}
\usage{
ler_documento_stj(diretorio = ".", arquivos = NULL,
  formato = c("html", "pdf"))
}
\arguments{
\item{diretorio}{informar diretório, apenas se
não informar arquivos.}

\item{arquivos}{se arquivos for NULL,
informar diretório.}

\item{formato}{"html" ou "pdf"}
}
\value{
tibble com sequenciais e documentos
}
\description{
Ler documentos do STJ
}
\examples{
\dontrun{
df <- ler_documento_pdf(diretorio = ".")
}
}
