% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_julgados_stj.R
\name{ler_julgados_stj}
\alias{ler_julgados_stj}
\title{Lê julgados do STJ}
\usage{
ler_julgados_stj(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{lista de arquivos}

\item{diretorio}{se arquivos não forem informados,
informar diretório}
}
\value{
dataframe
}
\description{
Lê julgados do STJ
}
\examples{
\dontrun{
df <- ler_julgados_stj()
}
}
