% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_julgados_stj2.R
\name{ler_julgados_stj2}
\alias{ler_julgados_stj2}
\title{Lê julgados do STJ, incluindo todos os campos}
\usage{
ler_julgados_stj2(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{lista de arquivos}

\item{diretorio}{se arquivos não forem informados,
informar diretório}
}
\value{
dataframe
}
\description{
Lê julgados do STJ, incluindo todos os campos
}
\examples{
\dontrun{
df <- ler_julgados_stj2()
}
}
