% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_metadocs_stj.R
\name{ler_metadocs_stj}
\alias{ler_metadocs_stj}
\title{Ler metadados do STJ}
\usage{
ler_metadocs_stj(arquivos = NULL, diretorio = ".",
  plano = "sequential")
}
\arguments{
\item{arquivos}{vetor com caminhos para arquivos}

\item{diretorio}{se arquivos for NULL, default para
diretório atual}

\item{plano}{ver `future::plan`}
}
\value{
tibble com números dos processos, metadado dos docs
    e sequencial para ser usado na função baixar_pdf_stj
}
\description{
Ler metadados do STJ
}
\examples{
\dontrun{
ler_metadocs_stj(diretorio=".")
}
}
