% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizar_docs_stj.R
\name{organizar_docs_stj}
\alias{organizar_docs_stj}
\title{Organiza todos os documentos decisivos, colocando-os numa única observação}
\usage{
organizar_docs_stj(metadocs, docs)
}
\arguments{
\item{metadocs}{retorno da função ler_metadocs_stj}

\item{docs}{retorno da função ler_documentos_stj}
}
\value{
tibble
}
\description{
Organiza todos os documentos decisivos, colocando-os numa única observação
}
\details{
Se você quer manter documentos que não são decisórios,
    tais como certidões, não use esta função.
}
\examples{
\dontrun{
df <- organizar_docs_stj(metadocs, docs)
}
}
