% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizar_stj.R
\name{organizar_stj}
\alias{organizar_stj}
\title{Organiza dados do stj a partir das bases julgados e detalhes}
\usage{
organizar_stj(detalhes, julgados)
}
\arguments{
\item{detalhes}{dados obtidos com ler_datalhes_stj}

\item{julgados}{dados obtidos com ler_julgados_stj}
}
\value{
dados organizadps
}
\description{
Organiza dados do stj a partir das bases julgados e detalhes
}
\examples{
\dontrun{
df <- organizar_stj(detalhes, julgados)
}
}
