% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-stj.R
\name{parse_stj}
\alias{parse_stj}
\title{Pega andamentos do STJ}
\usage{
parse_stj(arqs)
}
\arguments{
\item{arqs}{character vector com nomes dos arquivos html
baixados usando \code{\link{download_stj}}}
}
\description{
Pega andamentos do STJ a partir de arquivos já baixados.
}

