% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_stj_columns.R
\name{psql_columns}
\alias{psql_columns}
\title{psql_columns}
\usage{
psql_columns(con, tbl = NULL)
}
\arguments{
\item{con}{connection}

\item{tbl}{table name}
}
\value{
a data.frame with 18 columns describing the table columns.
}
\description{
psql_columns
}
