% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_stj_insert.R
\name{psql_stj_insert}
\alias{psql_stj_insert}
\title{Insere novos registros a uma tabela existente}
\usage{
psql_stj_insert(con, tbl = NULL, data = NULL, batch_size = 50000)
}
\arguments{
\item{con}{conexao}

\item{tbl}{tabela existente}

\item{data}{data.frame}

\item{batch_size}{number of rows to be inserted at a time.
Defaults to 50.000}
}
\value{
NULL se inserido corretamente
}
\description{
Insere novos registros a uma tabela existente
}
\examples{
\dontrun{
con <- dbx::dbxConnect()
DBI::dbCreateTable(con, "consumidor", consumidor)
psql_insert(con, "consumidor", consumidor)
}
}
