% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_stj_query.R
\name{psql_stj_query}
\alias{psql_stj_query}
\title{PostgreSQL full-text search}
\usage{
psql_stj_query(con, tbl, query = "", assuntos = NULL, classes = NULL,
  registros = NULL, origem = NULL, start = "2009-01-01",
  end = NULL)
}
\arguments{
\item{con}{connection}

\item{tbl}{table}

\item{query}{palavras a serem buscadas}

\item{assuntos}{Vetor de assuntos}

\item{classes}{vetor de classes}

\item{registros}{vetor de registros}

\item{origem}{vetor de origem}

\item{start}{data inicial no formato "yyyy-mm-dd"}

\item{end}{data final no formato "yyyy-mm-dd"}
}
\value{
tibble
}
\description{
PostgreSQL full-text search
}
\examples{
\dontrun{
dplyr::copy_to(con, "julgados", df)
df <- psql_stj_query(con, "julgados", "IRPF")
}
}
