% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramo.R
\docType{data}
\name{ramo}
\alias{ramo}
\title{ramo}
\format{
dataframe com 15 obs e duas colunas:
\describe{
  \item{\code{ramo}}{character descrição do ramo}
  \item{\code{sigla}}{character sigla do ramo}
}
}
\source{
\url{https://processo.stj.jus.br/processo/pesquisa/?aplicacao=processos.ea}
}
\usage{
ramo
}
\description{
Ramos do direito para usar nas funções stj_baixar_por...
}
\keyword{datasets}
