% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenciais.R
\name{sequenciais}
\alias{sequenciais}
\title{Extrai os números sequenciais dos htmls obtidos com baixar_processo_trf1.}
\usage{
sequenciais(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Se nulo, informar o diretório}

\item{diretorio}{Informar se os arquivos não foram informados}
}
\value{
tibble
}
\description{
Extrai os números sequenciais dos htmls obtidos com baixar_processo_trf1.
}
\examples{
\dontrun{
df <- sequenciais()
}
}
