% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj.R
\docType{package}
\name{stj-package}
\alias{stj}
\alias{stj-package}
\title{\code{stj} package}
\description{
Baixa  e organiza decisões do STJ
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jjesusfilho/stj}
  \item Report bugs at \url{https://github.com/jjesusfilho/stj/issues}
}

}
\author{
\strong{Maintainer}: Jose de Jesus Filho \email{jose@consudata.com.br} (\href{https://orcid.org/0000-0001-7415-7383}{ORCID})

}
