% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_acordaos_cjsg.R
\name{stj_baixar_acordao_cjsg}
\alias{stj_baixar_acordao_cjsg}
\title{Baixa ac처rd찾os do STJ a partir das urls lidas do stj_ler_cjsg}
\usage{
stj_baixar_acordao_cjsg(url, diretorio = ".")
}
\arguments{
\item{url}{Vetor de urls}

\item{diretorio}{Onde armazenar os arquivos}
}
\value{
pdfs
}
\description{
Baixa ac처rd찾os do STJ a partir das urls lidas do stj_ler_cjsg
}
