% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_cjsg.R
\name{stj_baixar_cjsg}
\alias{stj_baixar_cjsg}
\title{Extrair metadados da jurisprudência do STJ}
\usage{
stj_baixar_cjsg(
  livre = "",
  aspas = FALSE,
  data_inicial = "",
  data_final = "",
  base = "ACOR",
  n = NULL,
  diretorio = "."
)
}
\arguments{
\item{livre}{Se deixar em branco, informe as datas.}

\item{aspas}{Colocar entre aspas a busca}

\item{data_inicial}{Data inicial da publicação no formato "dd/mm/aaaa"}

\item{data_final}{Data final da publicação no formato "dd/mm/aaaa"}

\item{base}{"ACOR" ou "MONO"}

\item{n}{Número de páginas a serem baixadas. Se NULL,
baixa todas as páginas disponíveis}

\item{diretorio}{Informar onde baixar os htmls}
}
\value{
tibble com metadados
}
\description{
Extrair metadados da jurisprudência do STJ
}
\examples{
\dontrun{
    stj_baixar_cjsg(
   data_inicial = "01/07/2019",
    data_final = "30/07/2019"
    )
}
}
