% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_classe_processo.R
\name{stj_baixar_classe_processo}
\alias{stj_baixar_classe_processo}
\title{Baixar processo por número do STJ: classe + numero}
\usage{
stj_baixar_classe_processo(classe, numero, diretorio)
}
\arguments{
\item{classe}{Classe processsual}

\item{numero}{Número}

\item{diretorio}{Diretório onde armazenar os htmls}
}
\value{
html
}
\description{
Baixar processo por número do STJ: classe + numero
}
\examples{
\dontrun{
stj_baixar_classe_processo(classe = "hc", numero = "200000")
}
}
