% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_decisoes.R
\name{stj_baixar_decisoes}
\alias{stj_baixar_decisoes}
\title{Baixar decis천es das urls obtidas com stj_listar_decisoes}
\usage{
stj_baixar_decisoes(df, diretorio = ".")
}
\arguments{
\item{df}{tibble com url, registro e sequencial}

\item{diretorio}{Diret처rio onde colocar as decis천es}
}
\value{
pdfs
}
\description{
Baixar decis천es das urls obtidas com stj_listar_decisoes
}
