% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_decisoes2.R
\name{stj_baixar_decisoes2}
\alias{stj_baixar_decisoes2}
\title{Baixa decisões do stj com base no registro  e no sequencial.}
\usage{
stj_baixar_decisoes2(registro, sequencial, diretorio = ".")
}
\arguments{
\item{registro}{Número do registro}

\item{sequencial}{Número do sequencial}

\item{diretorio}{Caminho}
}
\value{
pdfs
}
\description{
Baixa decisões do stj com base no registro  e no sequencial.
}
