% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_docs_pdf.R
\name{stj_baixar_docs_pdf}
\alias{stj_baixar_docs_pdf}
\title{Baixa docs em pdf}
\usage{
stj_baixar_docs_pdf(sequencial, registro, data, diretorio = ".")
}
\arguments{
\item{sequencial}{Sequencial do documento}

\item{registro}{Número do registro}

\item{data}{Data no formato dd/mm/aaaa}

\item{diretorio}{Diretório onde armazenar os pdfs}
}
\value{
pdf
}
\description{
Baixa docs em pdf
}
\examples{
\dontrun{
stj_baixar_docs_pdf("112282908", "20200165980","04/08/2020")

}
}
