% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_documentos_pdf.R
\name{stj_baixar_documentos_pdf}
\alias{stj_baixar_documentos_pdf}
\title{Baixa documents a partir do tibble criado com ler_metacods_stj}
\usage{
stj_baixar_documentos_pdf(df, diretorio = ".")
}
\arguments{
\item{df}{Tibble}

\item{diretorio}{Diretório onde armazenar os arquivos}
}
\value{
pdfs
}
\description{
Baixa documents a partir do tibble criado com ler_metacods_stj
}
