% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_intimacoes.R
\name{stj_baixar_intimacoes}
\alias{stj_baixar_intimacoes}
\title{Baixa intimações do usuário especificado}
\usage{
stj_baixar_intimacoes(usuario = NULL, senha = NULL, diretorio = ".")
}
\arguments{
\item{usuario}{CPF do usuário. Se não for informado, irá procurar
pela variável de ambiente STJUSUARIO. Se não encontrar,
irá abrir uma caixa de diálogo para preenchimento.}

\item{senha}{Senha. Se não for informado, irá procurar
pela variável de ambiente STJSENHA. Se não encontrar,
irá abrir uma caixa de diálogo para preenchimento.}

\item{diretorio}{Diretório onde baixar o arquivo xls.}
}
\value{
Arquivo xls
}
\description{
Baixa intimações do usuário especificado
}
