% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_por_publicacao.R
\name{stj_baixar_por_publicacao}
\alias{stj_baixar_por_publicacao}
\title{Baixa consulta de processos do STJ por data da publicação}
\usage{
stj_baixar_por_publicacao(
  data_inicial,
  data_final,
  origem = "",
  ramo = "",
  n = NULL,
  diretorio = "."
)
}
\arguments{
\item{data_inicial}{Data inicial no formato dd/mm/aaaa.}

\item{data_final}{Data final no formato dd/mm/aaaa.}

\item{origem}{Sigla do(s) estado(s) de origem.}

\item{ramo}{Sigla do ramo do direito. Verifique o ramo
por meio do dataset incluso ramo.}

\item{n}{Número de páginas.}

\item{diretorio}{Diretório. Se não informado, atual.}
}
\value{
Html
}
\description{
Baixa consulta de processos do STJ por data da publicação
}
\examples{
\dontrun{
stj_baixar_por_publicacao(data_inicial = "01/01/2019",
                          data_final = "30/01/2019",
                          origem = c("SP", "AC"))
}

}
