% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_processo.R
\name{stj_baixar_processo}
\alias{stj_baixar_processo}
\title{Baixa processos do stj pelo número do processo ou do registro}
\usage{
stj_baixar_processo(numero = NULL, diretorio = ".", documentos = FALSE)
}
\arguments{
\item{numero}{Número do processo ou do registro}

\item{diretorio}{Diretório. Default para atual}

\item{documentos}{Se TRUE, baixa também documentos}
}
\value{
html baixado no diretório indicado
}
\description{
Você pode informar tanto o número do processo,
    quanto o número do registro.
}
\examples{
\dontrun{
baixar_processo_stj("2016/0074331-0")
baixar_processo_stj("01673163020133000000")
baixar_processo_stj("0167316-30.2013.3.00.0000")

}
}
