% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_processo_data.R
\name{stj_baixar_processo_data}
\alias{stj_baixar_processo_data}
\title{Baixa processos por data}
\usage{
stj_baixar_processo_data(dt_inicio = NULL, dt_fim = NULL, diretorio = ".")
}
\arguments{
\item{dt_inicio}{Data inicial no formato dd/mm/aaaa. Se não informar,
baixa os processos de hoje.}

\item{dt_fim}{Data final no formato dd/mm/aaaa.
Se não informar,  baixa os processos de hoje.}

\item{diretorio}{Onde armazenar os htmls}
}
\value{
html
}
\description{
Baixa processos por data
}
