% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_processo_originario.R
\name{stj_baixar_processo_originario}
\alias{stj_baixar_processo_originario}
\title{Baixa processos do stj pelo número originário do processo do processo ou do registro}
\usage{
stj_baixar_processo_originario(
  numero = NULL,
  diretorio = ".",
  documentos = FALSE
)
}
\arguments{
\item{numero}{Número do processo originário, ou seja, do TJ de origem}

\item{diretorio}{Diretório. Default para atual}

\item{documentos}{Se TRUE, baixa também documentos}
}
\value{
html baixado no diretório indicado
}
\description{
Você pode informar tanto o número do processo,
    quanto o número do registro.
}
