% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_publicacoes_dje.R
\name{stj_baixar_publicacoes_dje}
\alias{stj_baixar_publicacoes_dje}
\title{Baixa publicacoes do dje com base no registro e na data}
\usage{
stj_baixar_publicacoes_dje(registro, data, diretorio)
}
\arguments{
\item{registro}{Número do registro}

\item{data}{DAta no formato dd/mm/yyyy}

\item{diretorio}{Onde baixar os pdfs}
}
\value{
pdfs
}
\description{
Baixa publicacoes do dje com base no registro e na data
}
