% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_classificar_recurso2.R
\name{stj_classificar_recurso2}
\alias{stj_classificar_recurso2}
\title{Classifica recurso com apenas vetor de julgados}
\usage{
stj_classificar_recurso2(x)
}
\arguments{
\item{x}{Julgado ou dispositivo}
}
\value{
classificacao (string)
}
\description{
Classifica recurso com apenas vetor de julgados
}
\details{
Esta função se diferencia da outra porque sua entrada
    é um vetor e não um data.frame. Mais tarde, irei convertê-las
    em métodos de um mesmo genérico.
}
