% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_classificar_writs.R
\name{stj_classificar_writ}
\alias{stj_classificar_writ}
\title{Classificar writs(HC, MS, revisão criminal)}
\usage{
stj_classificar_writ(x, preprocessado = TRUE)
}
\arguments{
\item{x}{Texto da decisão}

\item{preprocessado}{Padrão para TRUE.  Se FALSE irá
remover espaços extras, colocar em caixa baixa e
remover acentos.}
}
\value{
classificação
}
\description{
Classificar writs(HC, MS, revisão criminal)
}
