% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_extrair_dispositivo.R
\name{stj_extrair_dispositivo}
\alias{stj_extrair_dispositivo}
\title{Extrai dispositivo da decisão}
\usage{
stj_extrair_dispositivo(x)
}
\arguments{
\item{x}{Inteiro teor da decisão}
}
\value{
character
}
\description{
Extrai dispositivo da decisão
}
