% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_extrair_papel.R
\name{stj_extrair_papel}
\alias{stj_extrair_papel}
\title{Extrai papeln da parte  no texto da decisão}
\usage{
stj_extrair_papel(x, padrao)
}
\arguments{
\item{x}{Inteiro teor da decisão}

\item{padrao}{Padrão, parte buscaa,depois de dois pontos e
espaço. Já ignora caixa.}
}
\value{
Papel da parte buscada
}
\description{
Extrai papeln da parte  no texto da decisão
}
