% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_extrair_partes_detalhes.R
\name{stj_extrair_partes_detalhes}
\alias{stj_extrair_partes_detalhes}
\title{Extrai partes ou somente detalhes da tibble detalhes}
\usage{
stj_extrair_partes_detalhes(df = NULL, dados = c("partes", "detalhes"))
}
\arguments{
\item{df}{tibble criada a partir da função ler_detalhes_stj}

\item{dados}{informar se "partes" ou "detalhes'
Default: "partes"}
}
\value{
tibble
}
\description{
Extrai partes ou somente detalhes da tibble detalhes
}
\examples{
\dontrun{
partes <- extrair_partes_detalhes(df,dados="partes")
}
}
