% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_cjsg_acor.R
\name{stj_ler_cjsg_acor}
\alias{stj_ler_cjsg_acor}
\title{Lê metadados de decisões colegiadas do STJ}
\usage{
stj_ler_cjsg_acor(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Se não informar arquivos, informar diretório}
}
\value{
Tibble
}
\description{
Lê metadados de decisões colegiadas do STJ
}
