% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_cjsg_mono.R
\name{stj_ler_cjsg_mono}
\alias{stj_ler_cjsg_mono}
\title{Lê decisões monocráticas do STJ}
\usage{
stj_ler_cjsg_mono(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Se não informar arquivos, informar diretório}
}
\value{
Tibble
}
\description{
Lê decisões monocráticas do STJ
}
