% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_decisoes.R
\name{stj_ler_decisoes}
\alias{stj_ler_decisoes}
\title{Ler decisões baixadas com stj_baixar_decisoes}
\usage{
stj_ler_decisoes(arquivos = NULL, diretorio = ".", assinatura = FALSE)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Diretório se não informar arquivos}

\item{assinatura}{Manter assinatura no pé da página?}
}
\value{
Tibble
}
\description{
Ler decisões baixadas com stj_baixar_decisoes
}
