% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_decisoes_dje.R
\name{stj_ler_decisoes_dje}
\alias{stj_ler_decisoes_dje}
\title{Lê decisoes baixadas do DJE}
\usage{
stj_ler_decisoes_dje(arquivos = NULL, assinatura = FALSE, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{assinatura}{FALSE para remover.}

\item{diretorio}{Alternativa aos arquivos.}
}
\value{
tibble
}
\description{
Lê decisoes baixadas do DJE
}
