% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_detalhes.R
\name{stj_ler_detalhes}
\alias{stj_ler_detalhes}
\title{Ler detalhes dos processos do STJ}
\usage{
stj_ler_detalhes(arquivos = NULL, diretorio = ".", wide = FALSE)
}
\arguments{
\item{arquivos}{Se os caminhos para os arquivos forem fornecidos
o diretório é ignorado.}

\item{diretorio}{Diretório se arquivos não forem informados}

\item{wide}{Colocar no formato largo? Padrão é longo.}
}
\value{
tibble com detalhes do processo
}
\description{
Ler detalhes dos processos do STJ
}
\examples{
\dontrun{
df <- ler_detalhes_stj(diretorio = ".")
}
}
