% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_docs_pdf.R
\name{stj_ler_docs_pdf}
\alias{stj_ler_docs_pdf}
\title{Lê pdfs baixados com stj_baixar_docs_pdf}
\usage{
stj_ler_docs_pdf(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar se não informar arquivos}
}
\value{
Tibble com os julgados
}
\description{
Lê pdfs baixados com stj_baixar_docs_pdf
}
