% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_documentos.R
\name{stj_ler_documentos}
\alias{stj_ler_documentos}
\title{Ler documentos do STJ}
\usage{
stj_ler_documentos(
  arquivos = NULL,
  diretorio = ".",
  formato = c("html", "pdf")
)
}
\arguments{
\item{arquivos}{se arquivos for NULL,
informar diretório.}

\item{diretorio}{informar diretório, apenas se
não informar arquivos.}

\item{formato}{"html" ou "pdf"}
}
\value{
tibble com sequenciais e documentos
}
\description{
Ler documentos do STJ
}
\examples{
\dontrun{
df <- stj_ler_documentos(diretorio = ".")
}
}
