% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_julgados.R
\name{stj_ler_julgados}
\alias{stj_ler_julgados}
\title{Lê julgados do STJ}
\usage{
stj_ler_julgados(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{lista de arquivos}

\item{diretorio}{se arquivos não forem informados,
informar diretório}
}
\value{
dataframe
}
\description{
Lê julgados do STJ
}
\examples{
\dontrun{
df <- stj_ler_julgados()
}
}
