% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_julgados2.R
\name{stj_ler_julgados2}
\alias{stj_ler_julgados2}
\title{Lê julgados do STJ, incluindo todos os campos}
\usage{
stj_ler_julgados2(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{lista de arquivos}

\item{diretorio}{se arquivos não forem informados,
informar diretório}
}
\value{
dataframe
}
\description{
Lê julgados do STJ, incluindo todos os campos
}
\examples{
\dontrun{
df <- stj_ler_julgados2()
}
}
