% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_metadocs_classe.R
\name{stj_ler_metadocs_classe}
\alias{stj_ler_metadocs_classe}
\title{Ler metadados do STJ}
\usage{
stj_ler_metadocs_classe(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{vetor com caminhos para arquivos}

\item{diretorio}{se arquivos for NULL, default para
diretório atual}
}
\value{
tibble com números dos processos, metadado dos docs
    e sequencial para ser usado na função baixar_pdf_stj
}
\description{
Ler metadados do STJ
}
\examples{
\dontrun{
stj_ler_metadocs(diretorio=".")
}
}
