% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_por_autuacao.R
\name{stj_ler_por_autuacao}
\alias{stj_ler_por_autuacao}
\title{Lê arquivos baixados com stj_baixar_por_autuacao}
\usage{
stj_ler_por_autuacao(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de csvs}

\item{diretorio}{Informar diretorio se omitir arquivos}
}
\value{
Tibble
}
\description{
Lê arquivos baixados com stj_baixar_por_autuacao
}
