% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_processo_data.R
\name{stj_ler_processo_data}
\alias{stj_ler_processo_data}
\title{Lê processos baixados com stj_baixar_processo_data}
\usage{
stj_ler_processo_data(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de caminhos para os arquivos}

\item{diretorio}{Atlernativamente, informar diretório}
}
\value{
tibble
}
\description{
Lê processos baixados com stj_baixar_processo_data
}
