% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_listar_decisoes.R
\name{stj_listar_decisoes}
\alias{stj_listar_decisoes}
\title{Lê urls da aba decisões}
\usage{
stj_listar_decisoes(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos.}

\item{diretorio}{Informar diretório se não informar arquivos.}
}
\value{
tibble
}
\description{
Lê urls da aba decisões
}
