% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_organizar.R
\name{stj_organizar}
\alias{stj_organizar}
\title{Organiza dados do stj a partir das bases julgados e detalhes}
\usage{
stj_organizar(detalhes, julgados)
}
\arguments{
\item{detalhes}{dados obtidos com ler_datalhes_stj}

\item{julgados}{dados obtidos com ler_julgados_stj}
}
\value{
dados organizadps
}
\description{
Organiza dados do stj a partir das bases julgados e detalhes
}
\examples{
\dontrun{
df <- stj_organizar(detalhes, julgados)
}
}
