% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_organizar_docs.R
\name{stj_organizar_docs}
\alias{stj_organizar_docs}
\title{organizar documentos}
\usage{
stj_organizar_docs(
  metadocs = NULL,
  docs = NULL,
  plano = "sequencial",
  workers = 1
)
}
\arguments{
\item{metadocs}{tibble lida com ler_metadocs_stj}

\item{docs}{docs}

\item{plano}{usar somente se for via bash}

\item{workers}{usar somente se for via bash}
}
\value{
tibble
}
\description{
organizar documentos
}
