% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_sequenciais.R
\name{stj_sequenciais}
\alias{stj_sequenciais}
\title{Extrai os números sequenciais dos htmls obtidos com baixar_processo_stj.}
\usage{
stj_sequenciais(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Se nulo, informar o diretório}

\item{diretorio}{Informar se os arquivos não foram informados}
}
\value{
tibble
}
\description{
Extrai os números sequenciais dos htmls obtidos com baixar_processo_stj.
}
\examples{
\dontrun{
df <- stj_sequenciais()
}
}
