% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_julgados_stj.R
\name{baixar_julgados_stj}
\alias{baixar_julgados_stj}
\title{Baixar decisões do STJ}
\usage{
baixar_julgados_stj(
  livre = "",
  operador = "e",
  aspas = FALSE,
  repo = c("ACOR", "SUMU", "DTXT", "INFJ"),
  data_inicial = "",
  data_final = "",
  diretorio = "."
)
}
\arguments{
\item{livre}{campo livre}

\item{operador}{"e" ou "adj", padrão para "e"}

\item{aspas}{colocar entre aspas? padrão para TRUE}

\item{repo}{informar o repositóriorio, padrão para "ACOR"}

\item{data_inicial}{data no formato "dd/mm/yyyy"}

\item{data_final}{data no formato "dd/mm/yyyy"}

\item{diretorio}{padrão para diretorio atual}
}
\value{
htmls
}
\description{
Baixar decisões do STJ
}
\examples{
\dontrun{
baixar_julgados_stj(livre= "dissolu\u00e7\u00e3o irregular")
}
}
