% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizar_docs_stj.R
\name{organizar_docs_stj}
\alias{organizar_docs_stj}
\title{organizar documentos}
\usage{
organizar_docs_stj(metadocs = NULL, docs = NULL,
  plano = "sequencial", workers = 1)
}
\arguments{
\item{metadocs}{tibble lida com ler_metadocs_stj}

\item{docs}{docs}

\item{plano}{usar somente se for via bash}

\item{workers}{usar somente se for via bash}
}
\value{
tibble
}
\description{
organizar documentos
}
