% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizar_docs_stj.R
\name{organizar_docs_stj}
\alias{organizar_docs_stj}
\title{Organiza todos os documentos decisórios}
\usage{
organizar_docs_stj(metadocs = NULL, docs = NULL,
  plano = "sequencial", workers = 1)
}
\arguments{
\item{metadocs}{Resposta da função ler_metadocs_stj}

\item{docs}{Resposta da função ler_documentos_stj}

\item{plano}{Use plano 'multicore' ou 'multiprocess' apenas
se for rodar via bash para paralelizar}

\item{workers}{Se for paralelizar, apenas via bash,
o número de processos deve ser menor que o
número de cores. Essa operação usa 100% de cada cpu}
}
\value{
tibble
}
\description{
Organiza todos os documentos decisórios
}
\details{
Se você quer manter documentos que não são decisórios,
    tais como certidões, não use esta função.
}
\examples{
\dontrun{
df <- organizar_docs_stj(metadocs, docs)
}

}
