% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_baixar_cjsg.R
\name{stj_baixar_cjsg}
\alias{stj_baixar_cjsg}
\title{Extrair metadados da jurisprudência do STJ}
\usage{
stj_baixar_cjsg(
  livre = "",
  aspas = FALSE,
  dt_ini = "",
  dt_fim = "",
  dt_ini_pb = "",
  dt_fim_pb = "",
  base = "ACOR",
  n = NULL,
  diretorio = "."
)
}
\arguments{
\item{livre}{Se deixar em branco, informe as datas.}

\item{aspas}{Colocar entre aspas a busca}

\item{dt_ini}{Data inicial do julgamento no formato dd/mm/aaaa}

\item{dt_fim}{Data final do julgamento no formato dd/mm/aaaa}

\item{dt_ini_pb}{Data inicial da publicação no formato "dd/mm/aaaa"}

\item{dt_fim_pb}{Data final da publicação no formato "dd/mm/aaaa"}

\item{base}{"ACOR" ou "MONO"}

\item{n}{Número de páginas a serem baixadas. Se NULL,
baixa todas as páginas disponíveis}

\item{diretorio}{Informar onde baixar os htmls}
}
\value{
tibble com metadados
}
\description{
Extrair metadados da jurisprudência do STJ
}
\examples{
\dontrun{
    stj_baixar_cjsg(
   dt_ini = "01/07/2019",
    dt_fim = "30/07/2019"
    )
}
}
