% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_decisoes_dje.R
\name{stj_ler_decisoes_dje}
\alias{stj_ler_decisoes_dje}
\title{Ler decisões do STJ baixados do dje}
\usage{
stj_ler_decisoes_dje(arquivos = NULL, diretorio = ".", assinatura = FALSE)
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Alternativamente, informar diretórios.}

\item{assinatura}{FALSE para remover as assinaturas.}
}
\value{
tibble
}
\description{
Ler decisões do STJ baixados do dje
}
