% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stj_ler_fases.R
\name{stj_ler_fases}
\alias{stj_ler_fases}
\title{Ler fases dos processos do STJ}
\usage{
stj_ler_fases(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Se os caminhos para os arquivos forem fornecidos
o diretório é ignorado.}

\item{diretorio}{Diretório se arquivos não forem informados}
}
\value{
tibble com fases do processo
}
\description{
Ler fases dos processos do STJ
}
\examples{
\dontrun{
df <- ler_fases_stj(diretorio = ".")
}
}
